upload-image-service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
h11==0.14.0
idna==3.6
pillow==10.2.0
pydantic==2.6.2
pydantic_core==2.16.3
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
